+++ src/types/broadcast/dokument.ts
@@ -1,6 +1,10 @@
 import { EditDocumentConfig } from "../dokument";
 
 export interface EditDocumentBroadcastMessage {
+    documentFile?: string;
+    /**
+     * @deprecated Use documentFile instead
+     */
     document: string;
     config?: EditDocumentConfig;
 }