+++ src/types/broadcast/index.ts
@@ -1,2 +1,3 @@
 export * from "./broadcast";
+export * from "./dokument";
 export * from "./person";